Add InterpolatedSpreadDiscountCurve and PiecewiseSpreadYieldCurve
These classes allow bootstrapping a curve as a spread to another curve.
This is useful, for example, to build related curves when one of them
has fewer liquid benchmark instruments. Building as a spread allows the
curve with fewer benchmarks to inherit the shape from the curve with
more benchmarks.

Also, refactor LocalBootstrap's PenaltyFunction so that it does not
pollute global namespace and does not need to be a friend of
PiecewiseYieldCurve.

Also, auto-enable extrapolation for ForwardSpreadedTermStructure based
on the original curve for consistency with ZeroSpreadedTermStructure.